requirements.txt
dotenv==0.9.9
Faker==37.11.0
mysql-connector==2.2.9
mysql-connector-python==9.4.0
python-dotenv==1.1.1
tzdata==2025.2